Move ChecksummedBytes into a new module
