+++ mountpoint-s3/src/upload.rs
@@ -9,6 +9,8 @@ use mountpoint_s3_crt::checksums::crc32c::{Crc32c, Hasher};
 use thiserror::Error;
 use tracing::error;
 
+use crate::checksums::combine_checksums;
+
 type PutRequestError<Client> = ObjectClientError<PutObjectError, <Client as ObjectClient>::ClientError>;
 
 const MAX_S3_MULTIPART_UPLOAD_PARTS: usize = 10000;
@@ -180,13 +182,6 @@ fn verify_checksums(review: UploadReview, expected_size: u64, expected_checksum:
     true
 }
 
-/// Calculates the combined checksum for `AB` where `prefix_crc` is the checksum for `A`,
-/// `suffix_crc` is the checksum for `B`, and `suffic_len` is the length of `B`.
-fn combine_checksums(prefix_crc: Crc32c, suffix_crc: Crc32c, suffix_len: usize) -> Crc32c {
-    let combined = ::crc32c::crc32c_combine(prefix_crc.value(), suffix_crc.value(), suffix_len);
-    Crc32c::new(combined)
-}
-
 #[cfg(test)]
 mod tests {
     use std::collections::HashMap;
@@ -196,7 +191,6 @@ mod tests {
         failure_client::countdown_failure_client,
         mock_client::{MockClient, MockClientConfig, MockClientError},
     };
-    use mountpoint_s3_crt::checksums::crc32c;
     use test_case::test_case;
 
     #[tokio::test]
@@ -340,16 +334,5 @@ mod tests {
 
         assert!(!client.contains_key(key));
         assert!(!client.is_upload_in_progress(key));
-    }
-
-    #[test]
-    fn test_combine_checksums() {
-        let buf: &[u8] = b"123456789";
-        let (buf1, buf2) = buf.split_at(4);
-        let crc = crc32c::checksum(buf);
-        let crc1 = crc32c::checksum(buf1);
-        let crc2 = crc32c::checksum(buf2);
-        let combined = combine_checksums(crc1, crc2, buf2.len());
-        assert_eq!(combined, crc);
     }
 }
